backend/requirements.txt
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
firebase-admin==6.2.0
openai>=1.0.0
python-dotenv>=0.19.0